The ContextLabel was not updated for the "missed call" messagebox on qpe launch.
At the time ContextLabelManager tried to update the labels we did not have
an QApplication::activeWindow yet. So the text was not updated, there was no
signal/event which would trigger updating the the label. QApplication::setActiveWindow
gets called from qapplication_x11.cpp for XFocusIn. If we see this event coming
by call ContextLabelManager::updateContextLabels, assuming it starts a timer, and if
we have an active window the labels will be updated. Listening to QEvent::ApplicationActivated
does not work as this event is sent conditionally, listening on the WindowActivate was not
attempted, it felt wrong to do this in the eventFilter.