Revert "[gsm] Micro-optimasation to answer phonecalls faster"
The calypso modem disagrees on this. It is the case on wakeup and my tests back
then but now we do see issues when going to suspend and issueing commands. Revert
this and go back to testing another fix. This will add a small delay on answering
phone calls again.

This reverts commit c3da122d325107a8781abd655db99d0723667345.